requirements.txt
mss==9.0.1
openai==1.30.1
Pillow==10.3.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-cloud-vision==3.7.2
protobuf==5.26.1
PyGetWindow==0.0.9
pywin32==306
pywinauto==0.6.8
